requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
aiohttp==3.9.1
python-dotenv==1.0.0
numpy==1.24.3
python-dateutil==2.8.2
jinja2==3.1.2
python-multipart==0.0.6